requirements.txt
streamlit
easyocr
pytesseract
opencv-python
Pillow
pdf2image
transformers
torch
sentencepiece
markdownify
python-docx
python-dotenv
tqdm
requests
layoutparser[layoutmodels,tesseract]
llama-index
unstructured
ollama
